Refactors useCurrentLevelExecutionData to a provider
