+++ src/routes/PipelineRun/PipelineRun.test.tsx
@@ -58,7 +58,7 @@ vi.mock("@/services/executionService", () => ({
     isLoading: false,
     error: null,
     isFetching: false,
-    refetch: () => { },
+    refetch: () => {},
     enabled: false,
   }),
   countTaskStatuses: vi.fn(),
@@ -88,11 +88,11 @@ vi.mock("@/providers/ComponentSpecProvider", async (importOriginal) => {
 });
 
 vi.mock("@/hooks/useDocumentTitle", () => ({
-  useDocumentTitle: () => { },
+  useDocumentTitle: () => {},
 }));
 
 vi.mock("@/hooks/useFavicon", () => ({
-  useFavicon: () => { },
+  useFavicon: () => {},
 }));
 
 describe("<PipelineRun/>", () => {
